Modification eslint

+++ .eslintrc.js
@@ -18,7 +18,7 @@ module.exports = {
     ],
     "rules": {
         "semi": ["error", "always"],
-        "quotes": ["error", "double"]
+        "quotes": ["error", "single"]
     },
     "extends": "eslint:recommended"
 }
